0-binary_to_uint.c

+++ 0x14-bit_manipulation/100-get_endianness.c
@@ -5,7 +5,7 @@
  */
 int get_endianness(void)
 {
-	unsigned long char *ptr;
+	unsigned char *ptr;
 	unsigned long int hex = 0x1234;
 
 	ptr = (unsigned char *)&hex;
